Move helper methods to file CSharpDiagnosticAnalyzerQuickInfoProviderExtensions.cs
